% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{mobKriging-package}
\alias{mobKriging-package}
\title{Recursive Partitioning of Kriging Models}
\description{
Recursive partitioning of Kriging Models.
}
\details{
This package contains the ingredients needed to combine Kriging (\code{\link[DiceKriging]{km}} from package \pkg{DiceKriging})
with model-based recursive partitioning (\code{\link[party]{mob}} from package \pkg{party}).

\code{\link{kmModel}} is an object of class \code{\link[modeltools]{StatModel-class}} implemented in package \pkg{modeltools} that
provides an infra-structure for an unfitted \code{\link[DiceKriging]{km}} model.

Moreover, methods for \code{kmModel} objects (that contain an object of class \code{\link[DiceKriging]{km-class}})
for the generic functions
\code{\link[party]{reweight}}, \code{\link[stats]{deviance}}, \code{\link[sandwich]{estfun}}, and
\code{\link[stats]{predict}} are provided.
}
\references{
Zeileis, A., Hothorn, T. and Kornik, K. (2008), Model-based recursive partitioning.
\emph{Journal of Computational and Graphical Statistics}, \bold{17(2)} 492--514.

Roustant, O., Ginsbourger, D. and Deville, Y. (2012), DiceKriging, DiceOptim: Two R packages for the analysis of computer
experiments by Kriging-based metamodeling and optimization.
\emph{Journal of Statistical Software}, \bold{51(1)}, \url{http://www.jstatsoft.org/}.
}

