% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{node_diagnosticKriging}
\alias{node_diagnosticKriging}
\title{Panel Generating Functions for Partitioned Kriging Models}
\format{\preformatted{function (mobobj, id = TRUE, pop = TRUE, which = c("loo", "res", "qq"), 
    pointcol = "black", pointcex = 0.5, linecol = "black", linetype = 1, 
    ylines = 2, margins = rep(1.5, 4), kriging.type, trend.reestim = FALSE, 
    ...)  
 - attr(*, "class")= chr "grapcon_generator"
}}
\usage{
node_diagnosticKriging(mobobj, id = TRUE, pop = TRUE,
	which = c("loo", "res", "qq"),
	pointcol = "black", pointcex = 0.5, linecol = "black", linetype = 1,
	ylines = 2, margins = rep(1.5, 4),
	kriging.type, trend.reestim = FALSE,
	...)
}
\arguments{
\item{mobobj}{[\code{mob}] \cr
An object of class mob.}

\item{id}{[\code{logical(1)}] \cr
Should node IDs be plotted?}

\item{pop}{[\code{logical(1)}] \cr
Should the panel viewports be popped?}

\item{which}{[\code{character}] \cr
Which diagnostic plots should be drawn?
The default is to draw all three available plots, see details.}

\item{pointcol}{[\code{numeric} | \code{character}] \cr
Color for points.}

\item{pointcex}{[\code{numeric}] \cr
Character extension of points in scatter plots.}

\item{linecol}{[\code{numeric} | \code{character}] \cr
Color for lines.}

\item{linetype}{[\code{numeric} | \code{character}] \cr
Type for lines.}

\item{ylines}{[\code{numeric(1)}] \cr
Number of lines for spaces in y-direction.}

\item{margins}{[\code{numeric(4)}] \cr
Margins of the viewports.}

\item{kriging.type}{[\code{character(1)}] \cr
An optional character string corresponding to the kriging family, to be chosen between simple kriging (\code{"SK"}) or universal kriging (\code{"UK"}).}

\item{trend.reestim}{[\code{logical(1)}] \cr
Should the trend be reestimated when removing an observation? Defaults to \code{FALSE}.}

\item{\dots}{Additional arguments passed to callies.}
}
\description{
Some panel generating functions of class \code{grapcon_generator} (package \pkg{vcd}) for terminal nodes associated with Kriging models.
For more details see \code{\link[party]{node_bivplot}}.
}
\details{
\code{node_diagnosticKriging} draws diagnostic plots (see also \code{\link[DiceKriging]{plot.km}}).
There are three different plots available, based on the results of \code{\link[DiceKriging]{leaveOneOut.km}},
and which ones are drawn can be controlled via the \code{which} argument:
\describe{
   \item{\code{"loo"}}{plot of fitted values against response values,}
   \item{\code{"res"}}{plot of standardized residuals,}
   \item{\code{"qq"}}{qqplot of standardized residuals.}
}
The default is to draw all three diagnostic plots.
}
\examples{
if (require(DiceKriging)) {
    d <- 2L
    x <- seq(0, 1, length = 10L)
    design <- expand.grid(x1 = x, x2 = x)
    y <- apply(design, 1, branin)
    df <- data.frame(y = y, design)

    ## no trend (formula = ~ 1)
    m <- mobKriging(y ~ x1 + x2 | x1 + x2, data = df,
        km.args = list(control = list(trace = FALSE)),
        control = mob_control(objfun = deviance, minsplit = 20L, verbose = FALSE))

    ## default
    plot(m, terminal_panel = node_diagnosticKriging(m,
        linecol = "red", linetype = "dashed", kriging.type = "UK"))
    ## only leave-one-out plots
    plot(m, terminal_panel = node_diagnosticKriging(m, which = c("loo"),
        linecol = "red", linetype = "dashed", kriging.type = "UK"))
    ## only qq-plots
    plot(m, terminal_panel = node_diagnosticKriging(m, which = "qq",
        linecol = "grey", linetype = "dashed", kriging.type = "UK"))
    ## leave-one-out and residual plots
    plot(m, terminal_panel = node_diagnosticKriging(m, which = c("loo", "res"),
        linecol = "blue", linetype = "dashed", kriging.type = "UK"))
}
}
\keyword{datasets}

