% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.kmModel}
\alias{fitted.kmModel}
\alias{predict.kmModel}
\title{Predict a \code{\link{kmModel}}}
\usage{
\method{predict}{kmModel}(object, newdata = NULL, pred.type, ...,
  mob = TRUE)

\method{fitted}{kmModel}(object, pred.type, ..., mob = TRUE)
}
\arguments{
\item{object}{[\code{\link{kmModel}}] \cr
An object of class \code{"kmModel"} containing the fitted Kriging model.}

\item{newdata}{[\code{data.frame}] \cr
The cases to be predicted.}

\item{pred.type}{[\code{character(1)}] \cr
Argument \code{type} of \code{\link[DiceKriging]{predict.km}}.
A \code{character} string corresponding to the kriging family, to be chosen between simple kriging (\code{"SK"}) or universal kriging (\code{"UK"}).
(Called \code{pred.type} to avoid confusion with argument \code{type} of \code{\link[party]{predict.mob}}.)}

\item{\dots}{Further arguments to \code{\link[DiceKriging]{predict.km}}.}

\item{mob}{[\code{logical(1)}] \cr
   If \code{TRUE} the object returned by \code{\link[DiceKriging]{predict.km}} is a \code{data.frame} instead
   of a \code{list}.}
}
\value{
See \code{\link[DiceKriging]{predict.km}}.
A \code{list} (or \code{data.frame}) containing at least the Kriging mean and trend
computed at \code{newdata}.
}
\description{
Predict a Kriging model.
}
\references{
Roustant, O., Ginsbourger, D. and Deville, Y. (2012), DiceKriging, DiceOptim: Two R packages for the analysis of computer
experiments by Kriging-based metamodeling and optimization.
\emph{Journal of Statistical Software}, \bold{51(1)}, \url{http://www.jstatsoft.org/}.
}
\seealso{
\code{\link[modeltools]{Predict}}, \code{\link[stats]{predict}}, \code{\link[DiceKriging]{predict.km}}.
}

