% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predictMobKriging}
\alias{fittedMobKriging}
\alias{predictMobKriging}
\title{Predict a Partitioned Kriging Model}
\usage{
predictMobKriging(object, newdata = NULL, type = c("response", "node"),
  km.args = list())

fittedMobKriging(object, type = c("response", "node"), km.args = list())
}
\arguments{
\item{object}{[\code{\linkS4class{mob}}] \cr
A tree with fitted Kriging models in the terminal nodes.}

\item{newdata}{[\code{data.frame}] \cr
The cases to be predicted.}

\item{type}{[\code{character(1)}] \cr
Passed to \code{type} in \code{\link[party]{predict.mob}}: A character string specifying whether the response
or the ID of the associated terminal node should be predicted.}

\item{km.args}{[\code{list}] \cr
   Additional arguments to \code{\link[DiceKriging]{predict.km}}.}
}
\value{
If \code{type = "node"} a vector of associated terminal node IDs.

If \code{type = "response"} a \code{list} containing at least the Kriging mean and trend
computed at \code{newdata}, depending on \code{km.args}.
}
\description{
Interface function for predicting a partitioned Kriging model.
}
\details{
\code{predictMobKriging} calls \code{\link[party]{predict.mob}}, but handles name clashes and
generates prettier output.
}
\examples{
## currently nonsense example
if (require(DiceKriging)) {
    d <- 2L
    x <- seq(0, 1, length = 10L)
    design <- expand.grid(x1 = x, x2 = x)
    y <- apply(design, 1, branin)
    df <- data.frame(y = y, design)

    ## linear trend (formula = ~ x1 + x2)
    m <- mobKriging(y ~ x1 + x2 | x1 + x2, data = df,
        km.args = list(formula = ~ ., control = list(trace = FALSE)))

    ## predict the response
    pred <- predictMobKriging(m, newdata = design, km.args = list(type = "UK",
        se.compute = TRUE, cov.compute = TRUE))

    ## calculate fitted values, only trend and mean
    pred <- fittedMobKriging(m, km.args = list(type = "UK", se.compute = FALSE,
        light.return = TRUE))

    ## predict nodes
    nodes <- predictMobKriging(m, newdata = design, type = "node")
}
}
\seealso{
\code{\link[party]{predict.mob}}, \code{\link[DiceKriging]{predict.km}}, \code{\link{mobKriging}}.
}

