% Generated by roxygen2 (4.0.1): do not edit by hand
\name{reweight.kmModel}
\alias{deviance.kmModel}
\alias{estfun.kmModel}
\alias{reweight.kmModel}
\title{Methods for \code{\link{kmModel}} Objects}
\usage{
\method{reweight}{kmModel}(object, weights, ...)

\method{deviance}{kmModel}(object, ...)

\method{estfun}{kmModel}(x, ...)
}
\arguments{
\item{object}{[\code{\link{kmModel}}] \cr
An object of class \code{"kmModel"} containing a fitted Kriging model.}

\item{weights}{[\code{numeric}] \cr
A vector of observation weights.}

\item{\dots}{Further arguments.}

\item{x}{[\code{\link{kmModel}}] \cr
   An object of class \code{"kmModel"} containing a fitted Kriging model.}
}
\value{
\code{reweight}: The re-weighted fitted \code{"kmModel"} object.

\code{deviance}: The value of the deviance extracted from \code{object}.

\code{estfun}: The empirical estimating (or score) function, i.e., the derivatives of the log-likelihood with respect
  to the parameters, evaluated at the training data.
}
\description{
Methods for \code{\link{kmModel}} objects.
}
\seealso{
\code{\link[party]{reweight}}, \code{\link[stats]{deviance}}, \code{\link[sandwich]{estfun}},
\code{\link[DiceKriging]{km}}.
}

