% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{kmModel}
\alias{kmModel}
\title{\code{kmModel}}
\format{\preformatted{Formal class 'StatModel' [package "modeltools"] with 5 slots
  ..@ name        : chr "kriging model"
  ..@ dpp         :function (formula, data = list(), subset = NULL, na.action = NULL, frame = NULL, 
    enclos = sys.frame(sys.nframe()), other = list(), designMatrix = TRUE, 
    responseMatrix = TRUE, setHook = NULL, ...)  
  ..@ fit         :function (object, weights = NULL, noise.var = NULL, km.args = NULL, ...)  
  ..@ predict     :function (object, newdata = NULL, ...)  
  ..@ capabilities:Formal class 'StatModelCapabilities' [package "modeltools"] with 2 slots
  .. .. ..@ weights: logi FALSE
  .. .. ..@ subset : logi FALSE
}}
\usage{
kmModel
}
\description{
A class for an unfitted Kriging model.
}
\details{
\code{kmModel} is an object of class \code{\link[modeltools]{StatModel-class}} implemented in package \pkg{modeltools} that
provides an infra-structure for an unfitted \code{\link[DiceKriging]{km}} model.
}
\examples{
if (require(DiceKriging)) {
    d <- 2L
    x <- seq(0, 1, length = 4L)
    design <- expand.grid(x1 = x, x2 = x)
    y <- apply(design, 1, branin)
    df <- data.frame(y = y, design)

    ## data pre-processing
    mf <- dpp(kmModel, y ~ ., data = df)

    ## no trend (formula = ~ 1)
    m1 <- fit(kmModel, mf)
    ## equivalent
    m2 <- km(design = design, response = y)

    ## linear trend (formula = ~ x1 + x2)
    m1 <- fit(kmModel, mf, formula = ~ .)
    ## equivalent
    m2 <- km(formula = ~ ., design = design, response = y)

    ## predictions
    Predict(m1, pred.type = "UK")
    ## equivalent
    predict(m2, newdata = design, type = "UK")
}
}
\references{
Roustant, O., Ginsbourger, D. and Deville, Y. (2012), DiceKriging, DiceOptim: Two R packages for the analysis of computer
experiments by Kriging-based metamodeling and optimization.
\emph{Journal of Statistical Software}, \bold{51(1)}, \url{http://www.jstatsoft.org/}.
}
\seealso{
\code{\link[modeltools]{StatModel-class}}, \code{\link[DiceKriging]{km}}.
}
\keyword{datasets}

