% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{mobKriging-package}
\alias{mobKriging-package}
\title{Recursive Partitioning of Kriging Models}
\description{
This package provides an interface for model-based recursive partitioning (\code{\link[party]{mob}} from package \pkg{party}) of
Kriging models (\code{\link[DiceKriging]{km}} from package \pkg{DiceKriging}).
The main functions are \code{\link{mobKriging}} and \code{\link{predictMobKriging}}.
These allow to build segmented Kriging models by computing decision trees where each terminal node corresponds to a fitted Kriging model
and making predictions based on those models.
Partitioning of Kriging models can be particularly useful if the data contain numerical as well as categorical explanatory variables.
Since Kriging normally requires numerical explanatory variables the categorical variables cannot be taken into account directly but
can be used as partitioning variables in \code{\link{mobKriging}}.
Package \pkg{mobKriging} also provides different plots, among others diagnostic plots, for fitted segmented Kriging models.
}
\references{
Zeileis, A., Hothorn, T. and Kornik, K. (2008), Model-based recursive partitioning.
\emph{Journal of Computational and Graphical Statistics}, \bold{17(2)} 492--514.

Roustant, O., Ginsbourger, D. and Deville, Y. (2012), DiceKriging, DiceOptim: Two R packages for the analysis of computer
experiments by Kriging-based metamodeling and optimization.
\emph{Journal of Statistical Software}, \bold{51(1)}, \url{http://www.jstatsoft.org/}.
}

