% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predict.kmModel}
\alias{fitted.kmModel}
\alias{predict.kmModel}
\title{Predict a \code{\link{kmModel}}}
\usage{
\method{predict}{kmModel}(object, newdata = NULL, ...)

\method{fitted}{kmModel}(object, ...)
}
\arguments{
\item{object}{[\code{\link{kmModel}}] \cr
The fitted Kriging model.}

\item{newdata}{[\code{data.frame}] \cr
The cases to be predicted.}

\item{\dots}{Additional arguments to \code{\link[DiceKriging]{predict.km}}.}
}
\value{
See \code{\link[DiceKriging]{predict.km}}.
A \code{list} containing at least the Kriging mean and trend
computed at \code{newdata}.
}
\description{
Predict a fitted Kriging model.
}
\examples{
## We use the first example in the documentation of function km
if (require(DiceKriging)) {
    d <- 2L
    x <- seq(0, 1, length = 4L)
    design <- expand.grid(x1 = x, x2 = x)
    y <- apply(design, 1, branin)
    df <- data.frame(y = y, design)

    ## fitting the model
    mf <- dpp(kmModel, y ~ ., data = df)
    m <- fit(kmModel, mf, formula = ~ .)

    ## fitted values and prediction
    fitted(m, type = "UK")
    predict(m, type = "UK", se.compute = FALSE, light.return = TRUE)
}
}
\references{
Roustant, O., Ginsbourger, D. and Deville, Y. (2012), DiceKriging, DiceOptim: Two R packages for the analysis of computer
experiments by Kriging-based metamodeling and optimization.
\emph{Journal of Statistical Software}, \bold{51(1)}, \url{http://www.jstatsoft.org/}.
}
\seealso{
\code{\link[modeltools]{Predict}}, \code{\link[stats]{predict}}, \code{\link[DiceKriging]{predict.km}}, \code{\link{kmModel}}.
}

